requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
streamlit==1.37.1
openai==1.35.0
httpx==0.27.0
python-dotenv==1.0.0
